refactor: improve UseContext and update props
