+++ src/pages/UserFlex.tsx
@@ -1,7 +1,21 @@
-import React from "react";
+import { useContext } from "react";
+import Card from "../components/card/Card";
+import UserContext from "../utils/UserContext";
 
-export const UserFlex = (): JSX.Element => (
-    <header>
-        <h1>This is the UserFlex Page.</h1>
-    </header>
-)
+export const UserFlex = (): JSX.Element => {
+    const users = useContext(UserContext);
+
+    return (
+        <section>
+            <h1>This is the UserFlex Page.</h1>
+            {users.map((user) => (
+                <Card 
+                id={user.id}
+                url="image.jpg"
+                title={user.name}
+                description={user.description}
+                />
+            ))}
+        </section>
+    )
+}
