+++ src/utils/UserContext.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+type User = {
+    id: number
+    name: string
+    description: string
+  }
+
+const UserContext = React.createContext<User[]>([]);
+export default UserContext
